+++ masternode_helpers.py
@@ -20,9 +20,8 @@ class MasternodeHelpers:
                 self._node_address = str(net.node_address) # This is the same for all networks so don't write it to network config
                 net_config = self.get_network_config(network_name)
                 if net_config:
-                    net_fee_obj = NetFee(net)
                     self._active_networks_config[network_name] = net_config
-                    self._active_networks_config[network_name]['native_ticker'] = str(net_fee_obj.native_ticker)
+                    self._active_networks_config[network_name]['native_ticker'] = str(NetFee(net).native_ticker)
         except Exception as e:
             logger.error(f"Error fetching active networks: {e}")