+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
 ## 1.13
+
+### Changed
+- Add X-API-Key to Access-Control-Allow-Headers.